# services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TenHub テクニカルナレッジベースプラットフォームのAPI",
        "title": "TenHub API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/articles": {
            "get": {
                "description": "全ての記事を新しい順で返す",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wiki"
                ],
                "summary": "記事一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "新しい記事を作成する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wiki"
                ],
                "summary": "記事作成",
                "parameters": [
                    {
                        "description": "記事データ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "description": "指定したIDの記事を返す",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wiki"
                ],
                "summary": "記事詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "記事ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "指定したIDの記事を更新する（部分更新対応）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wiki"
                ],
                "summary": "記事更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "記事ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新データ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "指定したIDの記事を削除する",
                "tags": [
                    "wiki"
                ],
                "summary": "記事削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "記事ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "description": "全てのカテゴリを階層構造で返す",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "カテゴリ一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "新しいカテゴリを作成する（parent_id で階層化可能）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "カテゴリ作成",
                "parameters": [
                    {
                        "description": "カテゴリデータ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swagger.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "delete": {
                "description": "指定したIDのカテゴリを削除する",
                "tags": [
                    "categories"
                ],
                "summary": "カテゴリ削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "カテゴリID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "swagger.CreateArticleRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string",
                    "example": "1"
                },
                "content": {
                    "type": "string",
                    "example": "# Hello World\n本文です"
                },
                "title": {
                    "type": "string",
                    "example": "Go入門ガイド"
                }
            }
        },
        "swagger.CreateCategoryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Go"
                },
                "parent_id": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "swagger.UpdateArticleRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "更新内容"
                },
                "title": {
                    "type": "string",
                    "example": "更新タイトル"
                }
            }
        },
...
    }
}